feat: InstallmentRepository; service delegates to repo
